requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
Flask==1.0.2
gevent==1.3.6
greenlet==0.4.14
idna==2.7
itsdangerous==0.24
Jinja2==2.10
locust==0.0
locustio==0.9.0
lru-dict==1.1.6
MarkupSafe==1.0
more-itertools==4.3.0
msgpack==0.5.6
pathlib2==2.3.2
pluggy==0.7.1
psutil==5.4.7
py==1.6.0
pytest==3.7.4
pytest-flask==0.11.0
pyzmq==17.1.2
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
